AppSettingPathHelper created
