+++ PalladiumPayroll.Helper/AppSettingPathHelper.cs
@@ -1,0 +1,25 @@
+﻿using Microsoft.Extensions.Options;
+
+namespace PalladiumPayroll.Helper
+{
+    public class AppSettingPathHelper
+    {
+        private readonly DirectoryPathSetting _settings;
+
+        public AppSettingPathHelper(IOptions<DirectoryPathSetting> options)
+        {
+            _settings = options.Value;
+        }
+
+        public DirectoryPathSetting GetAppSettingDirectoryPath()
+        {
+            return _settings;
+        }
+    }
+
+    public class DirectoryPathSetting
+    {
+        public string EmployeeDocument { get; set; } = null!;
+    }
+
+}
