+++ PalladiumProject.DTOs/DTOs/Common/CommonResponseDTO.cs
@@ -1,25 +1,5 @@
-﻿using Microsoft.Extensions.Options;
-
-namespace PalladiumPayroll.DTOs.DTOs.Common
+﻿namespace PalladiumPayroll.DTOs.DTOs.Common
 {
-    public class AppSettingDirectoryPath
-    {
-        private readonly DirectoryPathSetting _settings;
-
-        public AppSettingDirectoryPath(IOptions<DirectoryPathSetting> options)
-        {
-            _settings = options.Value;
-        }
-
-        public DirectoryPathSetting GetAppSettingDirectoryPath()
-        {
-            return _settings;
-        }
-    }
-    public class DirectoryPathSetting
-    {
-        public string EmployeeDocument { get; set; } = null!;
-    }
 
     public class DropDownViewModel
     {
